requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.19.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.12.5
email-validator==2.3.0
python-multipart==0.0.20
